lesson 6: Components made, code clean, data segregated, used Named and default imports, added search bar thinking of user convenience
