+++ components/Footer.js
@@ -1,0 +1,7 @@
+const Footer = () => (
+  <>
+    <span>This is Footer</span>
+  </>
+);
+
+export default Footer;
